requirements.txt
bottle==0.12.13
marshmallow==3.12.1
peewee==2.10.1
PyJWT==1.5.2
python-dotenv==0.17.1
